JWT Request Authentication on MeetingController

+++ app/Http/Controllers/MeetingController.php
@@ -5,12 +5,15 @@ namespace App\Http\Controllers;
 use App\Meeting;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use JWTAuth;
 
 class MeetingController extends Controller
 {
     public function __construct()
     {
-        // $this->middleware('name');
+        $this->middleware('jwt.auth', ['only'=> [
+        'update', 'store', 'destroy'
+        ]]);
     }
 
     /**
@@ -47,13 +50,16 @@ class MeetingController extends Controller
             'title' => 'required',
             'description' => 'required',
             'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+//            'user_id' => 'required'
         ]);
 
+        if (! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=> 'User not found'], 404);
+        }
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
         $meeting = new Meeting([
             'time' => Carbon::createFromFormat('YmdHie', $time),
             'title' => $title,
@@ -113,18 +119,21 @@ class MeetingController extends Controller
             'title' => 'required',
             'description' => 'required',
             'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+//            'user_id' => 'required'
         ]);
+        if (! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=>'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
         $meeting = Meeting::with('users')->findOrFail($id);
         if(!$meeting->users()->where('users.id', $user_id)->first()){
             return response()->json(['msg'=>'User not registered for meeting, update not successful'], 401);
             }
-        $meeting->time = Carbon::createFromFomat('YmdHie', $time);
+        $meeting->time = Carbon::createFromFormat('YmdHie', $time);
         $meeting->title = $title;
         $meeting->description = $description;
         if (!$meeting->update()){
@@ -152,6 +161,12 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::findOrFail($id);
+        if (! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=>'User not found'], 404);
+        }
+        if (!$meeting->users()->where('users.id', $user->id)->first()){
+            return response()->json(['msg'=>'User not registered for the meeting, delete unsuccessful'], 401);
+        }
         $users = $meeting->users;
         $meeting->users()->detach();
         if (!$meeting->delete()){
